client: prevent locks from being acquired while closing
This commit fixes a race in Client.Close() that potentially allowed for
new locks to be acquired while closing / releasing all locks.

The race occurs when a call to AcquireLock() is in progress when Close()
is called and when AcquireLock() is called while closing, but before we
set Client.closed to true.